pyproject.toml
[tool.poetry]
name = "geneweaver-api"
version = "0.4.0a7"
description = "The Geneweaver API"
authors = [
    "Alexander Berger <alexander.berger@jax.org>",
    "Francisco Castellanos <francisco.castellanos@jax.org>",
    "Sejal Desai <sejal.desai@jax.org>"
]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://thejacksonlaboratory.github.io/geneweaver-docs/"
repository = "https://github.com/TheJacksonLaboratory/geneweaver-api"
packages = [
    {include = "geneweaver/api", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.9"

geneweaver-core = "^0.9.0a8"
fastapi = {extras = ["all"], version = "^0.99.1"}
uvicorn = {extras = ["standard"], version = "^0.24.0"}
geneweaver-db = "^0.3.0a16"
psycopg-pool = "^3.1.7"
requests = "^2.31.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
geneweaver-testing = "^0.1.2"
pytest-asyncio = "^0.21.0"

[tool.ruff]
select = ['F', 'E', 'W', 'A', 'C90', 'N', 'B', 'ANN', 'D', 'I', 'ERA', 'PD', 'NPY', 'PT']

[tool.ruff.per-file-ignores]
"tests/*" = ["ANN001", "ANN201"]
"src/*" = ["ANN101"]
"src/geneweaver/api/controller/*" = ["B008"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# TODO: Remove this when batch API is hooked up and tested.
[tool.coverage.report]
omit = ["*/batch.py"]